refactor(agent): split ACP tool status helpers
